added: support for es60
