+++ es60x/src/main/scala/tech/beshu/ror/es/ReadonlyRestPlugin.java
@@ -39,6 +39,7 @@ import org.elasticsearch.common.settings.Setting;
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.common.settings.SettingsFilter;
 import org.elasticsearch.common.util.BigArrays;
+import org.elasticsearch.common.util.PageCacheRecycler;
 import org.elasticsearch.common.util.concurrent.EsExecutors;
 import org.elasticsearch.common.util.concurrent.ThreadContext;
 import org.elasticsearch.common.xcontent.NamedXContentRegistry;
@@ -57,6 +58,7 @@ import org.elasticsearch.rest.RestController;
 import org.elasticsearch.rest.RestHandler;
 import org.elasticsearch.script.ScriptService;
 import org.elasticsearch.threadpool.ThreadPool;
+import org.elasticsearch.transport.Transport;
 import org.elasticsearch.transport.netty4.Netty4Utils;
 import org.elasticsearch.watcher.ResourceWatcherService;
 import scala.concurrent.duration.FiniteDuration;
@@ -67,7 +69,8 @@ import tech.beshu.ror.es.dlsfls.RoleIndexSearcherWrapper;
 import tech.beshu.ror.es.rradmin.RRAdminAction;
 import tech.beshu.ror.es.rradmin.TransportRRAdminAction;
 import tech.beshu.ror.es.rradmin.rest.RestRRAdminAction;
-import tech.beshu.ror.es.ssl.SSLTransportNetty4;
+import tech.beshu.ror.es.ssl.SSLNetty4HttpServerTransport;
+import tech.beshu.ror.es.ssl.SSLNetty4InternodeServerTransport;
 import tech.beshu.ror.es.utils.ThreadRepo;
 import tech.beshu.ror.utils.ScalaJavaHelper$;
 
@@ -140,6 +143,29 @@ public class ReadonlyRestPlugin extends Plugin
   }
 
   @Override
+  public Map<String, Supplier<Transport>> getTransports(Settings settings,
+                                                        ThreadPool threadPool,
+                                                        BigArrays bigArrays,
+                                                        CircuitBreakerService circuitBreakerService,
+                                                        NamedWriteableRegistry namedWriteableRegistry,
+                                                        NetworkService networkService) {
+    if(sslConfig.interNodeSsl().isDefined()) {
+      return Collections.singletonMap("ror_ssl_internode", () ->
+              new SSLNetty4InternodeServerTransport(
+                      settings,
+                      threadPool,
+                      networkService,
+                      bigArrays,
+                      namedWriteableRegistry,
+                      circuitBreakerService,
+                      sslConfig.interNodeSsl().get())
+      );
+    } else {
+      return Collections.EMPTY_MAP;
+    }
+  }
+
+  @Override
   public Map<String, Supplier<HttpServerTransport>> getHttpTransports(
       Settings settings,
       ThreadPool threadPool,
@@ -152,7 +178,7 @@ public class ReadonlyRestPlugin extends Plugin
     if(sslConfig.externalSsl().isDefined()) {
       return Collections.singletonMap(
           "ssl_netty4", () ->
-              new SSLTransportNetty4(
+              new SSLNetty4HttpServerTransport(
                   settings, networkService, bigArrays, threadPool, xContentRegistry, dispatcher, sslConfig.externalSsl().get()
               ));
     } else {
